step4/provisioner/requirements.txt
kubernetes==7.0.0
ruamel.yaml>0.15,<0.16
redis==2.10.6
